collision detection to prevent launched sprites in walls

+++ go-sprite.h
@@ -8,7 +8,7 @@ void go_reset();
 void go_controls();
 void go_line();
 
-void object_run_commands(uint8_t i);
+void object_run_commands(int i);
 
 #define GO_BREAK 0
 #define GO_NOT_MOVE 1
